factor out Languages Nav
