+++ .eslintrc.yml
@@ -20,7 +20,7 @@ rules:
   accessor-pairs: warn
   array-callback-return: warn
   block-scoped-var: warn
-  class-methods-use-this: warn
+  class-methods-use-this: off
   default-case: warn
   default-case-last: warn
   default-param-last: warn
